move import file meta detection to its own module
